move sample program to Samples folder
